Did some fixes, added better footprint for the flyback diodes:
Re-routed camera bus so that there is more ground for the crystal.
Added ground pour keepouts under the MCU.
Changed the footprint for the motor flyback diodes.
Changed the camera module to have 24 pins, 2 won't be connected
Checked the footprints for dsPIC, MCP and LV8045
Increased thermal pad for voltage regulator

Signed-off-by: Kolijn Wolfaardt <kolijn.wolfaardt@gmail.com>